backend/requirements.txt
chalice==1.31.1
boto3==1.28.85
SQLAlchemy==2.0.23
python-dotenv==1.0.0
PyJWT==2.8.0
requests==2.31.0
pydantic==1.10.13
pydantic[dotenv]==1.10.13
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
openai==1.3.7
